requirements.txt
selenium==4.36.0
python-dotenv
